src/requirements.txt
# Google Cloud Logging Integration
google-cloud-logging>=3.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
